# optimize upsample performance linear mode on CPU