--- a/aten/src/ATen/native/UpSampleNearest3d.cpp
+++ b/aten/src/ATen/native/UpSampleNearest3d.cpp
@@ -40,7 +40,7 @@ static void upsample_nearest3d_out_cpu_template(
       output_height,
       output_width);
 
-  output.resize_({nbatch, channels, output_depth, output_height, output_width});
+  output.resize_({nbatch, channels, output_depth, output_height, output_width}, input.suggest_memory_format());
   AT_ASSERT(
       input_depth > 0 && input_height > 0 && input_width > 0 &&
       output_depth > 0 && output_height > 0 && output_width > 0);