@mingfeima 2020-03-17
Follow up on #31452

`"cubic"` mode optimization and `channel last` memory format optimization (if necessary) will be handled in other PRs.

Benchmark is available from [op_bench-py](https://github.com/mingfeima/op_bench-py). The data is collected on Xeon Skylake 6148, 2*20 cores@2.4GHz.

To reproduce, for inference data: `./run.sh upsample_linear.py`; for training data: `./run.sh upsample_linear.py --train`

**Unit**: ms per each iteration, the lower the better.

#### single socket inference

operation | input size | original | this pr | speedup
-- | -- | -- | -- | --
upsample_nearest1d | [64, 9, 512] | 3.768 | 0.138 | 27.30
upsample_nearest2d | [8, 9, 64, 64] | 7.606 | 0.411 | 18.51
upsample_nearest3d | [8, 9, 32, 32, 32] | 291.662 | 9.355 | 31.18

#### single core inference

operation | input size | original | this pr | speedup
-- | -- | -- | -- | --
upsample_nearest1d | [64, 9, 512] | 3.688 | 2.354 | 1.57
upsample_nearest2d | [8, 9, 64, 64] | 7.631 | 6.548 | 1.17
upsample_nearest3d | [8, 9, 32, 32, 32] | 275.978 | 153.596 | 1.80

#### dual core training

operation | input size | original | this pr | speedup
-- | -- | -- | -- | --
upsample_nearest1d | [64, 9, 512] | 8.031 | 0.604 | 13.30
upsample_nearest2d | [8, 9, 64, 64] | 19.273 | 1.152 | 16.73
upsample_nearest3d | [8, 9, 32, 32, 32] | 639.977 | 22.860 | 28.00