@mingfeima 2020-04-28
#### update on nearest mode with channel last format
1. adjust grain size, the old grain size equals to `at::internal::GRAIN_SIZE * channels` which will prevent efficient usage of multiple cpu cores.
2. explicit vectorization

Test platform is `Xeon 8280 CPU 2*20 cores @ 2.70GHz`, benchmark [upsample_nearest.py](https://github.com/mingfeima/op_bench-py/blob/master/upsample_nearest.py)

Single socket receives **2.9x** speedup, single core receives **1.2x** speedup

```bash
(before, single socket)
upsample_nearest2d_cl: memory format:   nhwc, input size:  torch.Size([8, 9,   64, 64])
input.is_contiguous(memory_format=torch.channels_last):  True
input.is_contiguous():  False
forward time per iteration:   0.287 ms

(after, single socket)
upsample_nearest2d_cl:   memory format: nhwc, input size:    torch.Size([8, 9, 64, 64])
input.is_contiguous(memory_format=torch.channels_last):  True
input.is_contiguous():  False
forward time per iteration:   0.098 ms

(before, single core)
upsample_nearest2d_cl:   memory format: nhwc, input size:    torch.Size([8, 9, 64, 64])
input.is_contiguous(memory_format=torch.channels_last):  True
input.is_contiguous():  False
forward time per iteration:   1.770 ms

(after, single core)
upsample_nearest2d_cl:   memory format: nhwc, input size:    torch.Size([8, 9, 64, 64])
input.is_contiguous(memory_format=torch.channels_last):  True
input.is_contiguous():  False
forward time per iteration:   1.385 ms
```