@mingfeima 2020-04-28
#### update on linear mode with channel last format
Here also list performance of `nchw` to compare against `nhwc`. Actually `nhwc` should be treated as a turbo mode since it is more suitable for vectorization.

**comments:** 1. enlarge channel size to fully show the power of channel last format in the benchmark, vectorization is done on channel dim. 2. the old implementation has non-contiguous memory access pattern for nchw format, so the corresponding speedup ratio is even bigger than number of cores.

```bash
(before single socket)
upsample_bilinear:   memory format: nchw, input size:  [32,   128, 64, 64]
input.is_contiguous(memory_format=torch.channels_last):  False
input.is_contiguous():  True
forward time per iteration:   2260.420 ms

upsample_bilinear_cl: memory format:   nhwc, input size:  [32, 64, 64, 128]
input.is_contiguous(memory_format=torch.channels_last):  True
input.is_contiguous():  False
forward time per iteration:   2186.436 ms

(after single socket)
upsample_bilinear:   memory format: nchw, input size:  [32,   128, 64, 64]
input.is_contiguous(memory_format=torch.channels_last):  False
input.is_contiguous():  True
forward time per iteration:   19.432 ms

upsample_bilinear_cl: memory format:   nhwc, input size:  [32, 64, 64, 128]
input.is_contiguous(memory_format=torch.channels_last):  True
input.is_contiguous():  False
forward time per iteration:   5.859 ms

(before, single core)
upsample_bilinear:   memory format: nchw, input size:  [32,   128, 64, 64]
input.is_contiguous(memory_format=torch.channels_last):  False
input.is_contiguous():  True
forward time per iteration:   2321.760 ms

upsample_bilinear_cl: memory format:   nhwc, input size:  [32, 64, 64, 128]
input.is_contiguous(memory_format=torch.channels_last):  True
input.is_contiguous():  False
forward time per iteration:   2403.946 ms

(after, single core)
upsample_bilinear:   memory format: nchw, input size:  [32,   128, 64, 64]
input.is_contiguous(memory_format=torch.channels_last):  False
input.is_contiguous():  True
forward time per iteration:   365.771 ms

upsample_bilinear_cl: memory format:   nhwc, input size:  [32, 64, 64, 128]
input.is_contiguous(memory_format=torch.channels_last):  True
input.is_contiguous():  False
forward time per iteration:   72.455 ms