@mingfeima 2020-04-28
#### Updates
1. add `Channels Last` format kernel for `linear` mode, parallel on `num_batches` dim and vectorize on `channel` dim.
2. adjust grain size of `Channels Last` format kernel for `nearest` mode to `at::internal::GRAIN_SIZE / channels`, add explicit vectorization.

Test platform, Intel(R) Xeon(R) Gold 6248 CPU @ 2.50GHz, 2 * 20 cores. To reproduce, use [upsample_linear.py](https://github.com/mingfeima/op_bench-py/blob/master/upsample_linear.py)

#### upsample linear mode (unit: ms)
config | memory format | before | after | speedup
-- | -- | -- | -- | --
single socket (omp_num_threads=20) | NCHW | 2260.42 | 19.43 | 116.3
single socket (omp_num_threads=20) | NHWC | 2186.44 | 5.86 | 373.2
single core (omp_num_threads=1) | NCHW | 2321.76 | 365.77 | 6.3
single core (omp_num_threads=1) | NHWC | 2403.95 | 72.46 | 33.2

Input tensor size is enlarged to `[32, 128, 64, 64]` to fully show the performance diff between `nhwc` and `nchw` layout, (`nhwc` layout is easier for vectorization).